pyproject.toml
[project]
name = "spotify-thelist"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "spotipy>=2.25.1",
]
